requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1

livekit-agents[google]~=1.0
livekit-plugins-noise-cancellation~=0.2